# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simple CRUD",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/v1/user": {
            "get": {
                "tags": [
                    "Simple CRUD"
                ],
                "summary": "Get resources",
                "operationId": "App\\Http\\Controllers\\API\\UserController::index",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Keywords of search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "pagination": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Get resource of detail user",
                                                    "type": "array",
                                                    "items": {
                                                        "allOf": [
                                                            {
                                                                "$ref": "#/components/schemas/UserDetailResponse"
                                                            },
                                                            {
                                                                "allOf": [
                                                                    {
                                                                        "properties": {
                                                                            "user": {
                                                                                "$ref": "#/components/schemas/UserResponse"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                ]
                                                            }
                                                        ]
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Get resource of detail user",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/user/detail/{id}": {
            "get": {
                "tags": [
                    "Simple CRUD"
                ],
                "summary": "Get resources specific by id",
                "operationId": "App\\Http\\Controllers\\API\\UserController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user detail",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Get resource specific by id",
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/UserDetailResponse"
                                                        },
                                                        {
                                                            "allOf": [
                                                                {
                                                                    "properties": {
                                                                        "user": {
                                                                            "$ref": "#/components/schemas/UserResponse"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            ]
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Get resource specific by id",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Get resource specific by id",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/user/upsert": {
            "put": {
                "tags": [
                    "Simple CRUD"
                ],
                "summary": "Create new resource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDetailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserDetailRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/user/upsert/{id}": {
            "put": {
                "tags": [
                    "Simple CRUD"
                ],
                "summary": "Update resource specific by id",
                "operationId": "App\\Http\\Controllers\\API\\UserController::upsert",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user detail",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserDetailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Handle store or update",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Handle store or update",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserDetailRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserDetailResponse": {
                "title": "User Detail Response",
                "description": "User detail response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of user detail",
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status of user detail",
                        "type": "string",
                        "format": "string",
                        "example": "active"
                    },
                    "position": {
                        "title": "position",
                        "description": "Position of user detail",
                        "type": "string",
                        "format": "string",
                        "example": "Web Programmer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserDetailResponse"
                }
            },
...
            "Wrapper": {
                "title": "Wrapper Response",
                "description": "Wrapper Response",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data",
                        "type": "object",
                        "format": "object"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "message"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Wrapper"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}